Move fixtures to own file, fix long lines, de-asyncify tests and test for more metadata attributes,
